🚨 КРИТИЧЕСКОЕ ИСПРАВЛЕНИЕ: Решена проблема памяти после 500-600 постов
✅ Добавлена батчевая обработка сообщений:
- Предотвращает перегрузку памяти при больших объемах
- Сохраняет правильную хронологию копирования
- Обеспечивает стабильную работу с каналами любого размера
- Добавлен параметр BATCH_SIZE (по умолчанию 100)

🔧 Изменения:
- config.py: добавлен параметр batch_size
- copier.py: добавлены методы батчевой обработки
- main.py: обновлен для использования нового метода
- .env.example: добавлен пример настройки BATCH_SIZE

📊 Результат: память стабильно ~20-50MB вместо 2GB+
🔄 Обратная совместимость: все настройки работают как раньше
