+++ .env.example
@@ -1,20 +1,20 @@
 # Telegram API настройки (обязательные)
 # Получите на https://my.telegram.org/apps
-API_ID=your_api_id_here
-API_HASH=your_api_hash_here
-PHONE=+79802191710
+API_ID=12345678
+API_HASH=abcdef1234567890abcdef1234567890
+PHONE=+79123456789
 
 # Настройки групп (обязательные)
 # ID или username исходной группы/канала (откуда копируем)
-SOURCE_GROUP_ID=-1002399927446
+SOURCE_GROUP_ID=@source_group
 # ID или username целевой группы/канала (куда копируем)
-TARGET_GROUP_ID=-1003011027978
+TARGET_GROUP_ID=@target_group
 
 # Настройки поведения (опциональные)
 # Задержка между сообщениями в секундах
-DELAY_SECONDS=5
+DELAY_SECONDS=3
 # Максимум сообщений в час
-MESSAGES_PER_HOUR=120
+MESSAGES_PER_HOUR=30
 # Режим симуляции (true/false)
 DRY_RUN=false
 
@@ -23,6 +23,12 @@ DRY_RUN=false
 SESSION_NAME=telegram_copier
 # Файл для сохранения прогресса
 RESUME_FILE=last_message_id.txt
+
+# Отладка
+DEBUG_MESSAGE_IDS=false
+
+# ИСПРАВЛЕНИЕ: Батчевая обработка (предотвращает проблемы с памятью после 500-600 постов)
+BATCH_SIZE=100
 
 # Настройки прокси (опциональные)
 # PROXY_SERVER=proxy.example.com
