+++ copier.py
@@ -35,7 +35,7 @@ class TelegramCopier:
                  rate_limiter: RateLimiter, dry_run: bool = False, resume_file: str = 'last_message_id.txt',
                  use_message_tracker: bool = True, tracker_file: str = 'copied_messages.json', 
                  add_debug_tags: bool = False, flatten_structure: bool = False, 
-                 debug_message_ids: bool = False):
+                 debug_message_ids: bool = False, batch_size: int = 100):
         """
         Инициализация копировщика.
         
@@ -51,6 +51,7 @@ class TelegramCopier:
             add_debug_tags: Добавлять ли debug теги к сообщениям
             flatten_structure: Превращать ли вложенность в плоскую структуру (антивложенность)
             debug_message_ids: Добавлять ли ID сообщений к тексту для отладки
+            batch_size: Размер батча для обработки сообщений (по умолчанию 100)
         """
         self.client = client
         self.source_group_id = source_group_id
@@ -59,6 +60,10 @@ class TelegramCopier:
         self.dry_run = dry_run
         self.resume_file = resume_file
         self.logger = logging.getLogger('telegram_copier.copier')
+        
+        # ИСПРАВЛЕНИЕ: Размер батча для предотвращения проблем с памятью
+        self.batch_size = batch_size
+        self.logger.info(f"🔧 Инициализирован копировщик с размером батча: {batch_size}")
         
         # Кэш для entities
         self.source_entity = None
@@ -866,12 +871,295 @@ class TelegramCopier:
         
         return final_stats
     
-
+    async def copy_all_messages_batch(self, resume_from_id: Optional[int] = None) -> Dict[str, Any]:
+        """
+        ИСПРАВЛЕННЫЙ метод копирования всех сообщений с батчевой обработкой.
+        РЕШАЕТ ПРОБЛЕМУ ПАМЯТИ И ХРОНОЛОГИИ после 500-600 постов.
+        
+        Args:
+            resume_from_id: ID сообщения для возобновления с определенного места
+        
+        Returns:
+            Словарь со статистикой копирования
+        """
+        if not await self.initialize():
+            return {'error': 'Не удалось инициализировать группы/каналы'}
+        
+        # Инициализируем трекер с информацией о каналах
+        if self.message_tracker:
+            self.message_tracker.set_channels(
+                str(self.source_group_id), 
+                str(self.target_group_id)
+            )
+        
+        # Получаем общее количество сообщений
+        total_messages = await self.get_total_messages_count()
+        if total_messages == 0:
+            self.logger.warning("В исходной группе/канале нет сообщений")
+            return {'total_messages': 0, 'copied_messages': 0}
+        
+        self.logger.info(f"🔄 Начинаем БАТЧЕВОЕ копирование {total_messages} сообщений")
+        self.logger.info(f"📦 Размер батча: {self.batch_size} сообщений")
+        
+        # Определяем начальную позицию
+        if self.message_tracker and not resume_from_id:
+            # Используем трекер для определения последнего ID
+            last_copied_id = self.message_tracker.get_last_copied_id()
+            if last_copied_id:
+                resume_from_id = last_copied_id
+                self.logger.info(f"📊 Трекер: последний скопированный ID {last_copied_id}")
+        
+        min_id = resume_from_id if resume_from_id else 0
+        
+        # Инициализируем трекер прогресса
+        progress_tracker = ProgressTracker(total_messages)
+        
+        try:
+            # НОВАЯ АРХИТЕКТУРА: Батчевая обработка
+            batch_number = 1
+            
+            # Получаем батчи сообщений и обрабатываем их по порядку
+            async for batch in self._get_message_batches(min_id):
+                if not batch:  # Пустой батч - конец
+                    break
+                
+                self.logger.info(f"📦 Обрабатываем батч #{batch_number}: {len(batch)} сообщений")
+                
+                # Обрабатываем батч в хронологическом порядке
+                batch_stats = await self._process_message_batch(batch, progress_tracker)
+                
+                # Обновляем общую статистику
+                self.copied_messages += batch_stats['copied']
+                self.failed_messages += batch_stats['failed'] 
+                self.skipped_messages += batch_stats['skipped']
+                
+                # Логируем прогресс батча
+                self.logger.info(f"✅ Батч #{batch_number} завершен: "
+                               f"скопировано {batch_stats['copied']}, "
+                               f"ошибок {batch_stats['failed']}, "
+                               f"пропущено {batch_stats['skipped']}")
+                
+                # Сохраняем прогресс после каждого батча
+                if batch:
+                    last_message_id = max(msg.id for msg in batch)
+                    save_last_message_id(last_message_id, self.resume_file)
+                    self.logger.debug(f"💾 Прогресс сохранен: последний ID {last_message_id}")
+                
+                # Принудительная очистка памяти после каждого батча
+                await self._cleanup_batch_memory()
+                
+                batch_number += 1
+                
+                # Проверяем, достигли ли мы конца
+                if len(batch) < self.batch_size:
+                    self.logger.info("📄 Достигнут конец канала (батч меньше размера)")
+                    break
+        
+        except Exception as e:
+            self.logger.error(f"Критическая ошибка при батчевом копировании: {e}")
+            return {'error': str(e)}
+        
+        # Получаем финальную статистику
+        final_stats = progress_tracker.get_final_stats()
+        final_stats.update({
+            'copied_messages': self.copied_messages,
+            'failed_messages': self.failed_messages,
+            'skipped_messages': self.skipped_messages,
+            'batches_processed': batch_number - 1,
+            'batch_size': self.batch_size
+        })
+        
+        # Проверяем реальное количество сообщений в целевом канале
+        if self.copied_messages > 0 and not self.dry_run:
+            try:
+                target_count = await self.get_target_messages_count()
+                final_stats['target_messages_count'] = target_count
+                self.logger.info(f"🔍 Проверка: в целевом канале {target_count} сообщений")
+            except Exception as e:
+                self.logger.warning(f"Не удалось проверить целевой канал: {e}")
+        
+        self.logger.info(f"🎉 Батчевое копирование завершено! "
+                        f"Обработано {batch_number-1} батчей. "
+                        f"Скопировано: {self.copied_messages}, "
+                        f"Ошибок: {self.failed_messages}, "
+                        f"Пропущено: {self.skipped_messages}")
+        
+        return final_stats
     
-
+    async def _get_message_batches(self, min_id: int = 0):
+        """
+        Генератор батчей сообщений в хронологическом порядке.
+        КЛЮЧЕВОЕ ИСПРАВЛЕНИЕ: Не загружает все сообщения в память!
+        
+        Args:
+            min_id: Минимальный ID для начала обработки
+            
+        Yields:
+            Батчи сообщений размером self.batch_size
+        """
+        offset_id = 0  # Начинаем с самых новых для reverse=True
+        
+        while True:
+            try:
+                # Параметры для получения батча
+                iter_params = {
+                    'entity': self.source_entity,
+                    'limit': self.batch_size,
+                    'reverse': True,  # От старых к новым
+                }
+                
+                # Добавляем offset_id если не первый запрос
+                if offset_id > 0:
+                    iter_params['offset_id'] = offset_id
+                
+                # Добавляем min_id если возобновляем работу
+                if min_id > 0:
+                    iter_params['min_id'] = min_id
+                
+                # Получаем батч сообщений
+                batch = []
+                async for message in self.client.iter_messages(**iter_params):
+                    # Проверка дедупликации
+                    if self.deduplicator.is_message_processed(message):
+                        self.logger.debug(f"Сообщение {message.id} уже обработано, пропускаем")
+                        continue
+                    
+                    batch.append(message)
+                
+                # Если батч пустой - конец
+                if not batch:
+                    break
+                
+                # Обновляем offset_id для следующего батча
+                # В reverse=True режиме берем ID последнего (самого старого) сообщения
+                offset_id = batch[-1].id
+                
+                self.logger.debug(f"📦 Получен батч из {len(batch)} сообщений "
+                                f"(ID от {batch[0].id} до {batch[-1].id})")
+                
+                yield batch
+                
+                # Если батч меньше размера - это последний батч
+                if len(batch) < self.batch_size:
+                    break
+                    
+            except Exception as e:
+                self.logger.error(f"Ошибка получения батча сообщений: {e}")
+                break
     
-
+    async def _process_message_batch(self, batch: List[Message], progress_tracker: ProgressTracker) -> Dict[str, int]:
+        """
+        Обработка одного батча сообщений в хронологическом порядке.
+        
+        Args:
+            batch: Батч сообщений для обработки
+            progress_tracker: Трекер прогресса
+            
+        Returns:
+            Статистика обработки батча
+        """
+        batch_stats = {'copied': 0, 'failed': 0, 'skipped': 0}
+        
+        # Группируем альбомы в батче
+        albums = {}  # grouped_id -> список сообщений
+        single_messages = []
+        
+        for message in batch:
+            if hasattr(message, 'grouped_id') and message.grouped_id:
+                if message.grouped_id not in albums:
+                    albums[message.grouped_id] = []
+                albums[message.grouped_id].append(message)
+            else:
+                single_messages.append(message)
+        
+        # Обрабатываем сообщения в порядке их ID (хронология)
+        all_items = []
+        
+        # Добавляем одиночные сообщения
+        for msg in single_messages:
+            all_items.append(('single', msg))
+        
+        # Добавляем альбомы (берем первое сообщение альбома для сортировки)
+        for grouped_id, album_messages in albums.items():
+            album_messages.sort(key=lambda x: x.id)  # Сортируем внутри альбома
+            all_items.append(('album', album_messages))
+        
+        # Сортируем по ID для сохранения хронологии
+        all_items.sort(key=lambda item: item[1].id if item[0] == 'single' else item[1][0].id)
+        
+        # Обрабатываем каждый элемент
+        for item_type, item_data in all_items:
+            try:
+                success = False
+                if item_type == 'single':
+                    # Одиночное сообщение
+                    message = item_data
+                    success = await self.copy_single_message(message)
+                    
+                    if success:
+                        batch_stats['copied'] += 1
+                        save_last_message_id(message.id, self.resume_file)
+                    else:
+                        batch_stats['failed'] += 1
+                    
+                    progress_tracker.update(success)
+                
+                elif item_type == 'album':
+                    # Альбом
+                    album_messages = item_data
+                    success = await self.copy_album(album_messages)
+                    
+                    if success:
+                        batch_stats['copied'] += len(album_messages)
+                        last_album_id = max(msg.id for msg in album_messages)
+                        save_last_message_id(last_album_id, self.resume_file)
+                    else:
+                        batch_stats['failed'] += len(album_messages)
+                    
+                    for msg in album_messages:
+                        progress_tracker.update(success)
+                
+                # Соблюдаем лимиты скорости
+                if not self.dry_run and success:
+                    await self.rate_limiter.wait_if_needed()
+                    self.rate_limiter.record_message_sent()
+                
+            except FloodWaitError as e:
+                await handle_flood_wait(e, self.logger)
+                # Повторяем попытку для текущего элемента
+                if item_type == 'single':
+                    success = await self.copy_single_message(item_data)
+                    if success:
+                        batch_stats['copied'] += 1
+                    else:
+                        batch_stats['failed'] += 1
+                else:
+                    success = await self.copy_album(item_data)
+                    if success:
+                        batch_stats['copied'] += len(item_data)
+                    else:
+                        batch_stats['failed'] += len(item_data)
+                
+            except Exception as e:
+                self.logger.error(f"Ошибка обработки элемента {item_type}: {e}")
+                if item_type == 'single':
+                    batch_stats['failed'] += 1
+                else:
+                    batch_stats['failed'] += len(item_data)
+        
+        return batch_stats
     
+    async def _cleanup_batch_memory(self):
+        """Принудительная очистка памяти после обработки батча."""
+        import gc
+        gc.collect()  # Принудительный сбор мусора
+        
+        # Очищаем кэши дедупликатора (оставляем только последние записи)
+        self.deduplicator.cleanup_old_hashes(keep_recent=1000)
+        
+        # Небольшая пауза для системы
+        await asyncio.sleep(0.1)
+
     async def get_target_messages_count(self) -> int:
         """
         Получение количества сообщений в целевом канале для проверки.
